+++ src/main/groovy/life/qbic/datamodel/dtos/business/ProductItem.groovy
@@ -23,10 +23,38 @@ class ProductItem {
      */
     final Product product
 
+    /**
+     * Describes the total price of an item based on the quantity and unit price of a product
+     */
+    final double totalPrice
+
+    /**
+     * Describe the total discount price for a product (based on the quantity)
+     */
+    final double quantityDiscount
+
+    /**
+     *
+     * @param quantity The quantity of a product
+     * @param product The product for which an item is created
+     *
+     * @deprecated 2.11.0
+     */
+    @Deprecated
     ProductItem(double quantity, Product product) {
         this.quantity = quantity
         this.product = product
+        this.totalPrice = 0
+        this.quantityDiscount = 0
     }
 
-}
+    ProductItem(double quantity, Product product, double totalPrice, double quantityDiscount) {
+        this.quantity = quantity
+        this.product = product
+        this.totalPrice = totalPrice
+        this.quantityDiscount = quantityDiscount
+    }
 
+
+
+}